Progress Commit; QA Skeleton done.
A bare bones framework of how QA testing is to be done is complete. The
framework will allow product code to have checks that report to the test
suite, but not compile into the production binary. A separate project
using the same code will be create for the automated QA.